Move Menu component into separate file
